Refactored generateDlc.js to export functions as modules for testing Setup jest and wrote unit tests for functions in generateDlc
